pyproject.toml
[project]
name = "notegenie-five"
version = "0.1.0"
description = "This is the fith trial but first self code of NOTEGENIE, and app the summarize the note of students in seconds"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.123.5",
    "ipykernel>=7.1.0",
    "numpy>=2.3.5",
    "openrouter>=0.1.1",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "supabase>=2.25.0",
    "uvicorn>=0.38.0",
]
